refactor: add `DELETE` blog auth for exercise 4.21

+++ part4/bloglist_backend/controllers/blogs.js
@@ -42,8 +42,22 @@ blogsRouter.post('/', async (request, response) => {
 });
 
 blogsRouter.delete('/:id', async (request, response) => {
-  await Blog.findByIdAndDelete(request.params.id);
-  response.status(204).end();
+  const decodedToken = jwt.verify(request.token, process.env.SECRET);
+
+  if (!decodedToken.id) {
+    return response.status(401).json({ error: 'Invalid token' });
+  }
+
+  const user = await User.findById(decodedToken.id);
+  const blog = await Blog.findById(request.params.id);
+
+  if (blog.user.toString() === user._id.toString()) {
+    await Blog.findByIdAndDelete(request.params.id);
+    return response.status(204).end();
+  } else {
+    return response.status(401).json({ error: 'Invalid user' });
+  }
+
 });
 
 blogsRouter.put('/:id', async (request, response) => {
